fix(observability): close post-merge review gaps
