+++ configs/quality/test_governance_audit.yaml
@@ -803,13 +803,19 @@ oversized_test_module_inventory:
     source: tests/integration/test_grafana_config.py
     source_lines_after_split: 1561
     extracted_surface: tests/integration/test_grafana_surface_contracts.py
-    extracted_surface_lines: 697
+    extracted_surface_lines: 731
     owner: observability-dashboards
   - issue_ref: '#4505'
     source: tests/integration/test_grafana_surface_contracts.py
-    source_lines_after_split: 697
+    source_lines_after_split: 731
     extracted_surface: tests/integration/test_grafana_technical_policy.py
     extracted_surface_lines: 52
+    owner: observability-dashboards
+  - issue_ref: '#6326'
+    source: tests/unit/repo_backed/scripts/ops/observability/test_grafana_dashboard_tooling.py
+    source_lines_after_split: 983
+    extracted_surface: tests/unit/repo_backed/scripts/ops/observability/test_grafana_live_audit_tooling.py
+    extracted_surface_lines: 1234
     owner: observability-dashboards
   - issue_ref: '#4506'
     source: tests/unit/composition/runtime_builders/test_runner_builder_profiles.py
@@ -1047,6 +1053,9 @@ repo_backed_unit_test_exceptions:
   - path: tests/unit/repo_backed/scripts/ops/observability/test_grafana_dashboard_tooling.py
     target_lane: repo-backed-unit
     protected_surface: scripts/ops/{__main__.py,observability/grafana/**}
+  - path: tests/unit/repo_backed/scripts/ops/observability/test_grafana_live_audit_tooling.py
+    target_lane: repo-backed-unit
+    protected_surface: scripts/ops/observability/grafana/audit_live_grafana_panels.py
   - path: tests/unit/repo_backed/scripts/test_apply_entity_naming_rename_plan.py
     target_lane: repo-backed-unit
     protected_surface: src/tools/apply_entity_naming_rename_plan.py
